Update DiochanChanPerformer.java
The field "quote" in the request body is mandatory for posting on /b/: otherwise posting will fail and the user will get an "Invalid Server Response" error.
Other boards do not require this field and automatically assume "quote" is set to 0, causing all quotes to other users in a sent message to be considered as greentext. This can be easily tested by posting a message on a board that is not /b/ and quoting someone.
This behavior is useless and the front-end does not provide the option to disable quotes.
So the value should always be set to "1".